refactor(menuGroup): 별도 패키지 분리
