+++ src/main/java/kitchenpos/menu/domain/MenuCreatingValidator.java
@@ -7,7 +7,8 @@ import kitchenpos.core.domain.Quantity;
 import kitchenpos.core.exception.InvalidPriceException;
 import kitchenpos.menu.dto.MenuProductRequest;
 import kitchenpos.menu.dto.MenuRequest;
-import kitchenpos.menu.exception.NotFoundMenuGroupException;
+import kitchenpos.menugroup.exception.NotFoundMenuGroupException;
+import kitchenpos.menugroup.domain.MenuGroupRepository;
 import kitchenpos.product.domain.Product;
 import kitchenpos.product.domain.ProductRepository;
 import kitchenpos.product.exception.NotFoundProductException;
