+++ src/main/java/kitchenpos/menugroup/domain/MenuGroup.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu.domain;
+package kitchenpos.menugroup.domain;
 
 import kitchenpos.core.domain.Name;
