+++ src/test/java/kitchenpos/menugroup/application/MenuGroupServiceTest.java
@@ -1,8 +1,9 @@
-package kitchenpos.menu.application;
+package kitchenpos.menugroup.application;
 
 import kitchenpos.ServiceTest;
-import kitchenpos.menu.dto.MenuGroupRequest;
-import kitchenpos.menu.dto.MenuGroupResponse;
+import kitchenpos.menugroup.dto.MenuGroupRequest;
+import kitchenpos.menugroup.dto.MenuGroupResponse;
+import kitchenpos.menugroup.application.MenuGroupService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
